# Faster time_value.rb